chore: update dependencies and refactor imports for layers and user location contexts
