+++ package.json
@@ -32,6 +32,7 @@
 	},
 	"devDependencies": {
 		"@biomejs/biome": "2.0.0-beta.6",
+		"@tailwindcss/postcss": "^4.1.10",
 		"@tailwindcss/vite": "^4.1.10",
 		"@types/node": "^24.0.3",
 		"@types/react": "^18.3.5",
@@ -40,7 +41,7 @@
 		"autoprefixer": "^10.4.18",
 		"globals": "^15.9.0",
 		"lefthook": "1.11.13",
-		"postcss": "^8.4.35",
+		"postcss": "^8.5.6",
 		"tailwindcss": "^4.1.10",
 		"typescript": "^5.5.3",
 		"vite": "^5.4.2"
